Refactor inference modules with improved code organization and modularity
- Split generation.py into stage1.py and stage2.py for better code structure
- Move stereo and mono inference functions to dedicated modules
- Simplify import statements and reduce code complexity
- Remove redundant code and improve function signatures
- Enhance error handling and processing flexibility
- Prepare codebase for easier maintenance and future extensions
